Add compatibility shims, update portlibs & CI
Add lightweight shims and compatibility fixes so the project builds without full external SDKs: add mplayerwii shim, newlibcompat for __ctype_ptr__, URL escape/unescape and header, embeddata.sh and Makefile wiring to embed assets, and various include/API fixes (libwiigui, filebrowser, main/input). Update/modernize many vendored portlibs headers (Lua, mxml, libpng, 7-Zip) and add sevenzip Types forwarder. Disable old RAR implementation (stubbed). Improve CI: devkitPro install, portlibs selection, artifact verification and automated release job.

+++ external/libwiigui/gui_longtext.cpp
@@ -92,7 +92,7 @@ void GuiLongText::SetTextPos(int pos)
 
 	for(u32 i = 0; i < TextLines.size(); i++)
 	{
-		int curDiff = abs(TextLines[i] - pos);
+		int curDiff = abs((int)TextLines[i] - pos);
 		if(curDiff < diff)
 		{
 			diff = curDiff;
